fix: use once to prevent multiple callback calls
The aws api is unreliable with its errors. It's currently
possible for a call to succeed, call the callback, and then
later receive a connection error which is passed to the
already called callback. This results in callback called
twice errors. Leveraging once will prevent these errors from
the aws api.

License: MIT
Signed-off-by: Jacob Heun <jacobheun@gmail.com>